+++ client/src/Components/Admin/EditarImagenes.jsx
@@ -1,0 +1,23 @@
+
+
+
+export default function EditarImagenes(){
+
+    return(
+        <div>
+            <h3>Editar Imagenes</h3>
+            <ol>
+                {/* imagen.map(imagen => {
+                        return(
+                            <li key={imagen.id}>
+                                <img src={imagen.imagen
+                                <input type='file' name='imagen' value={imagen.imagen}/>
+                                <button id={imagen.id} onClick={(e) => {
+                                    dispatch(deleteImage(e.target.id));
+                                }}>Eliminar</button>
+                            </li>)
+                    })  */}
+            </ol>
+        </div>
+    )
+}
